refactor: remove save-fhir-data endpoint
