+++ containers/ecr-viewer/src/app/api/process-zip/types.ts
