--- containers/ecr-viewer/src/app/api/route.ts
@@ -1,9 +1,0 @@
-import { NextResponse } from "next/server";
-
-/**
- * Health check for ECR Viewer
- * @returns Response with status OK.
- */
-export async function GET() {
-  return NextResponse.json({ status: "OK" }, { status: 200 });
-}
